requirements.txt
torch
numpy
pycryptodome